# contevent/contevent-backend/src/main/java/com/contevent/backend/model/Company.java
package com.contevent.backend.model;

public class Company {
    private Long id;
    private String name;
    private String description;

    public Company() {}

    public Company(Long id, String name, String description) {
        this.id = id;
        this.name = name;
        this.description = description;
    }

    public Long getId() { return id; }
    public void setId(Long id) { this.id = id; }
    public String getName() { return name; }
    public void setName(String name) { this.name = name; }
    public String getDescription() { return description; }
    public void setDescription(String description) { this.description = description; }
}