# contevent/contevent-backend/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Contevent API
  description: |
    API for Contevent - an application that helps people connect
    through shared activities and events.
  version: 0.1.0
  contact:
    name: Contevent Team

servers:
  - url: http://localhost:8080
    description: Local development

paths:
  /events:
    get:
      operationId: listEvents
      summary: List all events
      tags: [Events]
      parameters:
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category
      responses:
        '200':
          description: A list of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
    post:
      operationId: createEvent
      summary: Create a new event
      tags: [Events]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventRequest'
      responses:
        '201':
          description: Event created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'

  /events/{eventId}:
    get:
      operationId: getEvent
      summary: Get event details
      tags: [Events]
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Event details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Event not found

  /events/{eventId}/applications:
    get:
      operationId: getApplications
      summary: List applications for an event
      tags: [Applications]
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: A list of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Application'
    post:
      operationId: applyToEvent
      summary: Apply to join an event
      tags: [Applications]
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApplicationRequest'
      responses:
        '201':
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'

  /events/{eventId}/applications/{applicationId}:
    put:
      operationId: reviewApplication
      summary: Accept or deny an application
      tags: [Applications]
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: applicationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewApplicationRequest'
      responses:
        '200':
          description: Application reviewed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'

  /groups:
    post:
      operationId: createGroup
      summary: Create a new group
      tags: [Groups]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
      responses:
        '201':
          description: Group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'

  /companies:
    post:
      operationId: createCompany
      summary: Register a new company
      tags: [Companies]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCompanyRequest'
      responses:
        '201':
          description: Company registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'

components:
  schemas:
    Event:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        category:
          type: string
        dateTime:
          type: string
          format: date-time
        location:
          type: string
        maxParticipants:
          type: integer
        creatorId:
          type: integer
          format: int64
        creatorType:
          type: string
          enum: [USER, GROUP, COMPANY]
        status:
          type: string
          enum: [OPEN, CLOSED, CANCELLED]

    CreateEventRequest:
      type: object
      required: [title, dateTime, maxParticipants]
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        dateTime:
          type: string
          format: date-time
        location:
          type: string
        maxParticipants:
          type: integer
        creatorId:
          type: integer
          format: int64
        creatorType:
          type: string
          enum: [USER, GROUP, COMPANY]

    Application:
      type: object
      properties:
        id:
          type: integer
          format: int64
        eventId:
          type: integer
          format: int64
        applicantId:
          type: integer
          format: int64
        status:
          type: string
          enum: [PENDING, ACCEPTED, DENIED]
        message:
          type: string

    CreateApplicationRequest:
      type: object
      properties:
        applicantId:
          type: integer
          format: int64
        message:
          type: string

    ReviewApplicationRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ACCEPTED, DENIED]

    Group:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        memberIds:
          type: array
          items:
            type: integer
            format: int64

    CreateGroupRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string

    Company:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string

    CreateCompanyRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string